Examiner updates
