+++ strr-examiner-web/app/enums/routes.ts
@@ -1,0 +1,4 @@
+export enum RoutesE {
+    DASHBOARD = '/dashboard',
+    EXAMINE = '/examine'
+  }
