+++ strr-examiner-web/app/pages/examine/index.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { RoutesE } from '~/enums/routes'
+import { useExaminerStore } from '~/store/examiner'
+
+const localePath = useLocalePath()
+const { getNextApplication } = useExaminerStore()
+
+const nextApp = await getNextApplication()
+navigateTo(localePath(`${RoutesE.EXAMINE}/${nextApp}`))
+
+</script>
